fixed EXC_BAD_ACCESS
- free_elms() will now free each node in pb_list
- fixed where ps->ops was not initialized
- fixed free()ing bug in get_ps()
- fixed seg_id management issues
- passes 2,3,4,5 tests

+++ src/checker/get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: eesaki <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/14 15:17:05 by eesaki            #+#    #+#             */
-/*   Updated: 2020/04/25 20:07:59 by eesaki           ###   ########.fr       */
+/*   Updated: 2020/08/03 06:25:56 by eesaki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int		get_line(char **str, char **line)
 
 	while ((nl = ft_strchr(*str, '\n')))
 	{
-		*line = ft_strndup(*str, (nl - *str));
+		*line = ft_strndup(*str, (size_t)(nl - *str));
 		tmp = ft_strdup(nl + 1);
 		free(*str);
 		*str = tmp;
